requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
arrow==1.2.3
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
blessed==1.20.0
cachetools==5.3.1
clean-fid==0.1.35
click==8.1.7
croniter==1.4.1
dateutils==0.6.12
deepdiff==6.5.0
diffusers==0.12.1
einops==0.6.1
exceptiongroup==1.1.3
fastapi==0.103.1
filelock==3.12.4
frozenlist==1.4.0
fsspec==2023.9.1
ftfy==6.1.1
google-auth==2.23.0
google-auth-oauthlib==1.0.0
grpcio==1.58.0
h11==0.14.0
huggingface-hub==0.17.1
imhist==0.0.4
importlib-metadata==6.8.0
imwatermark==0.0.2
inquirer==3.1.3
invisible-watermark==0.2.0
itsdangerous==2.1.2
jinja2==3.1.2
kornia==0.6.8
lightning==2.0.9
lightning-cloud==0.5.38
lightning-utilities==0.9.0
markdown==3.4.4
markdown-it-py==3.0.0
markupsafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
numpy==1.24.4
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.8.0.76
ordered-set==4.1.0
packaging==23.1
protobuf==4.24.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.1.1
pydantic-core==2.4.0
pydeprecate==0.3.2
pygments==2.16.1
pyjwt==2.8.0
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==1.7.7
pytz==2023.3.post1
pywavelets==1.4.1
pyyaml==6.0.1
readchar==4.0.5
regex==2023.8.8
requests-oauthlib==1.3.1
rich==13.5.3
rsa==4.9
scipy==1.10.1
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
starsessions==1.3.0
taming-transformers-rom1504==0.0.6
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tokenizers==0.12.1
torchmetrics==0.10.0
tqdm==4.66.1
traitlets==5.10.0
transformers==4.19.2
uvicorn==0.23.2
wcwidth==0.2.6
websocket-client==1.6.3
websockets==11.0.3
werkzeug==2.3.7
yarl==1.9.2
zipp==3.16.2
-e git+https://github.com/openai/CLIP.git@main#egg=clip
